refactor: improve terminal organization
- Split Terminal into smaller components (CommandInput, CommandOutput)
- Created custom hook for terminal logic (useTerminal)
- Extracted utilities and constants
- Improved code organization with proper folder structure
- Added index file for exports
