+++ apps/web/src/components/terminal/constants.ts
@@ -1,0 +1,24 @@
+export const INITIAL_COMMANDS = [
+  {
+    input: "npm start",
+    output: [
+      "Ready - started server on 0.0.0.0:3000",
+      "Event compiled successfully"
+    ]
+  }
+];
+
+export const COMMAND_SUGGESTIONS = [
+  'npm start',
+  'npm install',
+  'npm run build',
+  'npm run dev',
+  'git status',
+  'git add .',
+  'git commit -m',
+  'git push',
+  'git pull',
+  'git branch',
+  'git checkout',
+  'clear'
+];
