+++ apps/web/src/components/terminal/index.ts
@@ -1,0 +1,7 @@
+export { Terminal } from './Terminal';
+export { CommandInput } from './CommandInput';
+export { CommandOutput } from './CommandOutput';
+export { useTerminal } from './useTerminal';
+export * from './types';
+export * from './utils';
+export * from './constants';
